+++ src/components/button/button-test.stories.tsx
@@ -95,14 +95,14 @@ export const ButtonAsASiblingStory = ({
   ...args
 }: Partial<ButtonProps>) => {
   return (
-    <div>
+    <Box>
       <Button subtext={subtext} {...args} onClick={action("click")}>
         {children}
       </Button>
       <Button subtext={subtext} {...args} onClick={action("click")} ml={2}>
         {children}
       </Button>
-    </div>
+    </Box>
   );
 };
 
@@ -202,7 +202,7 @@ export const generateButtons = (
 export const generateButtonsNoWrapStory = (buttonType: ButtonTypes) => {
   return BUTTON_SIZES.map((size) => {
     return (
-      <div key={`${buttonType}-${size}`}>
+      <Box key={`${buttonType}-${size}`}>
         <Box width="100px">
           <Button buttonType={buttonType} noWrap size={size}>
             Long button text
@@ -236,7 +236,7 @@ export const generateButtonsNoWrapStory = (buttonType: ButtonTypes) => {
             Long button text
           </Button>
         </Box>
-      </div>
+      </Box>
     );
   });
 };
@@ -259,6 +259,18 @@ export const NoWrapTertiaryButtonsStory = () => {
 
 NoWrapTertiaryButtonsStory.storyName = "noWrap tertiary";
 
+export const NoWrapGradientWhiteButtonsStory = () => {
+  return <Box>{generateButtonsNoWrapStory("gradient-white")}</Box>;
+};
+
+NoWrapGradientWhiteButtonsStory.storyName = "noWrap gradient-white";
+
+export const NoWrapGradientGreyButtonsStory = () => {
+  return <Box>{generateButtonsNoWrapStory("gradient-grey")}</Box>;
+};
+
+NoWrapGradientGreyButtonsStory.storyName = "noWrap gradient-grey";
+
 export const NoWrapDashedButtonsStory = () => {
   return <Box>{generateButtonsNoWrapStory("dashed")}</Box>;
 };
@@ -278,7 +290,7 @@ export const IconOnlyButtonsStory = () => {
       {BUTTON_VARIANTS.map((buttonType) => {
         return BUTTON_SIZES.map((size) => {
           return (
-            <div key={`${buttonType}-${size}`}>
+            <Box key={`${buttonType}-${size}`}>
               <Box width="100px">
                 <Button
                   buttonType={buttonType}
@@ -287,7 +299,7 @@ export const IconOnlyButtonsStory = () => {
                   aria-label={binIcon}
                 />
               </Box>
-            </div>
+            </Box>
           );
         });
       })}
@@ -301,9 +313,9 @@ export const generateFullWidthButtonsStory = (buttonType: ButtonTypes) => {
   const props = { buttonType, fullWidth: true };
   return (
     <Box>
-      <div key={`${buttonType}-${buttonType}`}>
+      <Box key={`${buttonType}-${buttonType}`}>
         {BUTTON_SIZES.map((size) => (
-          <div key={`${buttonType}-${buttonType}-${size}`}>
+          <Box key={`${buttonType}-${buttonType}-${size}`}>
             <Button key="basic" size={size} {...props} ml={2}>
               {size}
             </Button>
@@ -354,10 +366,10 @@ export const generateFullWidthButtonsStory = (buttonType: ButtonTypes) => {
                 </Button>
               </Box>
             )}
-          </div>
+          </Box>
         ))}
         {BUTTON_SIZES.map((size) => (
-          <div key={`${buttonType}-${buttonType}-${size}-destructive`}>
+          <Box key={`${buttonType}-${buttonType}-${size}-destructive`}>
             <Button key="basic" size={size} destructive {...props} ml={2}>
               {size}
             </Button>
@@ -418,10 +430,10 @@ export const generateFullWidthButtonsStory = (buttonType: ButtonTypes) => {
                 </Button>
               </>
             )}
-          </div>
+          </Box>
         ))}
         {BUTTON_SIZES.map((size) => (
-          <div key={`${buttonType}-${buttonType}-${size}-disabled`}>
+          <Box key={`${buttonType}-${buttonType}-${size}-disabled`}>
             <Button key="basic" size={size} disabled {...props} ml={2}>
               {size}
             </Button>
@@ -482,9 +494,9 @@ export const generateFullWidthButtonsStory = (buttonType: ButtonTypes) => {
                 </Button>
               </>
             )}
-          </div>
+          </Box>
         ))}
-      </div>
+      </Box>
     </Box>
   );
 };
@@ -506,6 +518,18 @@ export const FullWidthTertiaryButtonsStory = () => {
 };
 
 FullWidthTertiaryButtonsStory.storyName = "fullWidth tertiary";
+
+export const FullWidthGradientWhiteButtonsStory = () => {
+  return <Box>{generateFullWidthButtonsStory("gradient-white")}</Box>;
+};
+
+FullWidthGradientWhiteButtonsStory.storyName = "fullWidth gradient-white";
+
+export const FullWidthGradientGreyButtonsStory = () => {
+  return <Box>{generateFullWidthButtonsStory("gradient-grey")}</Box>;
+};
+
+FullWidthGradientGreyButtonsStory.storyName = "fullWidth gradient-grey";
 
 export const FullWidthDashedButtonsStory = () => {
   return <Box>{generateFullWidthButtonsStory("dashed")}</Box>;
@@ -555,6 +579,30 @@ export const TertiaryButtonIconAfter = () => {
 
 TertiaryButtonIconAfter.storyName = "tertiary icon after";
 
+export const GradientWhiteButtonIconBefore = () => {
+  return <Box>{generateButtons("gradient-white", "before")}</Box>;
+};
+
+GradientWhiteButtonIconBefore.storyName = "gradient-white icon before";
+
+export const GradientWhiteButtonIconAfter = () => {
+  return <Box>{generateButtons("gradient-white", "after")}</Box>;
+};
+
+GradientWhiteButtonIconAfter.storyName = "gradient-white icon after";
+
+export const GradientGreyButtonIconBefore = () => {
+  return <Box>{generateButtons("gradient-grey", "before")}</Box>;
+};
+
+GradientGreyButtonIconBefore.storyName = "gradient-grey icon before";
+
+export const GradientGreyButtonIconAfter = () => {
+  return <Box>{generateButtons("gradient-grey", "after")}</Box>;
+};
+
+GradientGreyButtonIconAfter.storyName = "gradient-grey icon after";
+
 export const DashedButtonIconBefore = () => {
   return <Box>{generateButtons("dashed", "before")}</Box>;
 };
@@ -581,7 +629,7 @@ DarkBackgroundButtonIconAfter.storyName = "darkBackground icon after";
 
 export const ButtonDifferentTypes = (props: Partial<ButtonProps>) => {
   return (
-    <div>
+    <Box>
       <Button buttonType="primary" {...props}>
         Primary
       </Button>
@@ -594,6 +642,681 @@ export const ButtonDifferentTypes = (props: Partial<ButtonProps>) => {
       <Button buttonType="dashed" {...props}>
         Dashed
       </Button>
-    </div>
+      <Button buttonType="gradient-white" {...props}>
+        Gradient white
+      </Button>
+      <Button buttonType="gradient-grey" {...props}>
+        Gradient Grey
+      </Button>
+    </Box>
   );
 };
+
+export const DisabledButtonsWithImageChildren = () => (
+  <Box>
+    <Button buttonType="primary" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Primary
+    </Button>
+    <Button buttonType="secondary" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Secondary
+    </Button>
+    <Button buttonType="tertiary" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Tertiary
+    </Button>
+    <Button buttonType="dashed" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Dashed
+    </Button>
+    <Button buttonType="gradient-white" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Gradient white
+    </Button>
+    <Button buttonType="gradient-grey" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Gradient Grey
+    </Button>
+    <br />
+    <Button buttonType="primary" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Primary
+    </Button>
+    <Button buttonType="secondary" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Secondary
+    </Button>
+    <Button buttonType="tertiary" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Tertiary
+    </Button>
+    <Button buttonType="dashed" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Dashed
+    </Button>
+    <Button buttonType="gradient-white" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Gradient white
+    </Button>
+    <Button buttonType="gradient-grey" disabled>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        width="40"
+        height="40"
+        viewBox="0 0 120 100"
+      >
+        <defs>
+          <filter
+            id="a"
+            width="200%"
+            height="200%"
+            x="-50%"
+            y="-50%"
+            filterUnits="objectBoundingBox"
+          >
+            <feOffset
+              dx="1"
+              dy="4"
+              in="SourceAlpha"
+              result="shadowOffsetOuter1"
+            />
+            <feColorMatrix
+              values="0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0.1 0"
+              in="shadowOffsetOuter1"
+              result="shadowMatrixOuter1"
+            />
+            <feMerge>
+              <feMergeNode in="shadowMatrixOuter1" />
+              <feMergeNode in="SourceGraphic" />
+            </feMerge>
+          </filter>
+        </defs>
+        <g
+          fill="none"
+          fillRule="evenodd"
+          filter="url(#a)"
+          transform="translate(26 16)"
+        >
+          <path
+            fill="#255BC7"
+            d="M62.097.05c-1.99.4-41.597 36.6-41.597 36.6l10.15 9.745s31.447-41.57 32.64-43.16c1.196-1.592.797-3.58-1.193-3.184z"
+          />
+          <path
+            fill="#004089"
+            d="M19.738 37.643s-3.245 1.39-4.82 3.545c-1.7 2.328-1.782 6.37-1.782 6.37l6.322 6.594s3.154-.18 6.078-2.08c2.925-1.898 4.318-4.48 4.318-4.48l-10.116-9.95z"
+          />
+          <path
+            fill="#FFAB00"
+            d="M12.58 48.447s-6.61 0-9.388 4.115C.412 56.678 4.976 60.516 0 65c7.563-2.694 11.09 1.55 14.7-.97 3.61-2.52 4.208-8.81 4.208-8.81l-6.327-6.773z"
+          />
+        </g>
+      </svg>
+      Gradient Grey
+    </Button>
+  </Box>
+);
+
+DisabledButtonsWithImageChildren.storyName =
+  "disabled button with img children";